requirements.txt
# Core dependencies
chainlit==1.0.505
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3

# Together AI
openai==1.6.1  # Together AI compatible

# File processing
python-docx==1.1.0  # DOCX processing
Pillow==10.1.0  # Image processing

# Web scraping and search
requests==2.31.0
beautifulsoup4==4.12.2
serpapi==0.1.5

# Email
sendgrid==6.11.0

# Utilities
pandas==2.1.4
numpy==1.26.2
python-magic==0.4.27  # File type detection

# Logging
loguru==0.7.2

# Session management
redis==5.0.1  # Optional: for session storage

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0